requirements.txt
anitya_schema==2.3.0
asgiref==3.9.1
attrs==25.3.0
Automat==25.4.16
black==25.1.0
blinker==1.9.0
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.3
click==8.2.1
cloudinary==1.44.1
colorama==0.4.6
config==0.5.1
constantly==23.10.4
crispy-bootstrap4==2025.6
crochet==2.1.1
cryptography==45.0.6
decorator==5.2.1
defusedxml==0.7.1
Django==5.2.5
django-cloudinary-storage==0.3.0
django-crispy-forms==2.4
fedora-messaging-the-new-hotness-schema==1.2.1
fedora_messaging==3.8.0
flake8==7.3.0
gssapi==1.9.0
gunicorn==23.0.0
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
jsonschema==4.25.1
jsonschema-specifications==2025.4.1
koji==1.35.3
mccabe==0.7.0
mypy_extensions==1.1.0
packaging==25.0
pathspec==0.12.1
pika==1.3.2
pillow==11.3.0
platformdirs==4.3.8
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycodestyle==2.14.0
pycparser==2.22
pyflakes==3.4.0
PyJWT==2.10.1
pyOpenSSL==25.1.0
python-bugzilla==3.3.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
redis==5.3.1
referencing==0.36.2
requests==2.32.5
requests-gssapi==1.3.0
rpds-py==0.27.0
service-identity==24.2.0
setuptools==80.9.0
six==1.17.0
sqlparse==0.5.3
the-new-hotness==1.3.0
tomli==2.2.1
transliterate==1.10.2
Twisted==25.5.0
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
whitenoise==6.9.0
wrapt==1.17.3
zope.interface==7.2
